setting up the react router
