+++ front-end/delphi-web-app/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^6.28.0",
     "survey-creator-react": "^1.12.8",
     "survey-react-ui": "^1.12.8"
   },
